requirements.txt
matplotlib==3.9.1
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.1.105
PyYAML==6.0.1
torch==2.3.1
torchdata==0.7.1
torchdiffeq==0.2.4